requirements.txt
flask
requests
python-dotenv
numpy<2
docling
onnxruntime
opencv-python-headless
trafilatura
apscheduler
beautifulsoup4
